# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for conscientious citizens",
        "title": "Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login by login/email/phone + password. Returns a JWT access token and the authenticated user profile.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT access token and authenticated user",
                        "schema": {
                            "$ref": "#/definitions/response.AuthLogin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/maps/reload-cities": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reloads in-memory supported cities from zones table. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Reload supported cities cache",
                "responses": {
                    "200": {
                        "description": "Reload successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/maps/reverse": {
            "get": {
                "description": "Выполняет reverse geocoding: по координатам точки на карте возвращает человекочитаемый адрес.\nИспользуется, когда пользователь выбирает место на карте, а frontend должен показать адрес выбранной точки.\n\nСначала backend локально проверяет, попадает ли точка хотя бы в одну поддерживаемую зону проекта в PostGIS.\nЕсли точка не попадает ни в одну зону, возвращается ошибка. Если попадает, backend определяет адрес через локальный кеш и/или Nominatim/OpenStreetMap.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Получение адреса по координатам",
                "parameters": [
                    {
                        "type": "number",
                        "example": 53.2051714,
                        "description": "Широта точки",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "example": 50.1334676,
                        "description": "Долгота точки",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Найденный адрес по координатам",
                        "schema": {
                            "$ref": "#/definitions/response.ReverseGeocodeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные координаты или отсутствующие query-параметры",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Адрес по координатам не найден",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "422": {
                        "description": "Точка находится вне зоны работы проекта",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса карт",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/maps/search": {
            "get": {
                "description": "Выполняет поиск адреса по текстовому запросу пользователя через Nominatim/OpenStreetMap. Используется, когда пользователь вводит адрес вручную в строке поиска на карте.\n\nСервис может вернуть до 4 наиболее подходящих вариантов адреса. Если frontend передаёт query-параметр city, он трактуется как город пользователя и должен входить в список поддерживаемых городов проекта.\nЕсли в строке запроса уже указан поддерживаемый город, он имеет приоритет над query-параметром city. Если указан неподдерживаемый город или найдены только адреса вне поддерживаемых зон проекта, возвращается ошибка.\nЕсли по запросу не найдено ни одного подходящего адреса, возвращается ошибка 404.\n\nПримеры запросов:\n- \"Самара проспект Ленина 1\"\n- \"Самара Ленина 1\"\n- \"Ленина 1\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Поиск адреса по строке",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Ленина 1",
                        "example": "Ленина 1",
                        "description": "Строка поиска адреса",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "samara",
                        "example": "samara",
                        "description": "Город-подсказка пользователя. Для Swagger по умолчанию подставляется samara.",
                        "name": "city",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "До 4 наиболее подходящих адресов",
                        "schema": {
                            "$ref": "#/definitions/response.SearchAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос, например пустой или слишком короткий q",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Подходящих адресов не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "422": {
                        "description": "Указан неподдерживаемый город или найдены только адреса вне зоны работы проекта",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса карт",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.User": {
            "type": "object",
            "properties": {
                "apartment": {
                    "type": "string"
                },
                "avatar_url": {
                    "description": "AvatarURL stores identifier or URL of avatar image (e.g. S3 key).",
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "house": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_blocked": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "user, admin, premium",
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.Login": {
            "type": "object",
            "required": [
                "identifier",
                "password"
            ],
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "user123"
                },
                "password": {
                    "type": "string",
                    "example": "qwerty123"
                }
            }
        },
...
        "response.AuthLogin": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some informative error message"
                }
            }
        },
        "response.GeoAddress": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Самара"
                },
                "full_address": {
                    "type": "string",
                    "example": "1, проспект Ленина, Октябрьский район, Самара, городской округ Самара, Самарская область, Приволжский федеральный округ, 443110, Россия"
                },
                "house_number": {
                    "type": "string",
                    "example": "1"
                },
                "lat": {
                    "type": "number",
                    "example": 53.2051714
                },
                "lon": {
                    "type": "number",
                    "example": 50.1334676
                },
                "road": {
                    "type": "string",
                    "example": "проспект Ленина"
                }
            }
        },
...
        "response.ReverseGeocodeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.GeoAddress"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SearchAddressResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.GeoAddress"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token with Bearer prefix. Example: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}